rename the server/routers file to routes
