+++ server/index.js
@@ -3,7 +3,7 @@ import express from 'express';
 import passport from 'passport';
 import session from 'express-session';
 import { Strategy as LocalStrategy } from 'passport-local';
-import bcrypt from 'bcrypt'; 
+import bcrypt from 'bcrypt';
 import db from './config/db.js'; // database conection
 import env from 'dotenv'
 import flash from 'connect-flash'
@@ -22,14 +22,14 @@ const port = process.env.PORT
 const error505msg = "Sorry! It seems we have a problem with our servers. Please try again later."
 // Session configuration
 app.use(session({
-    secret: process.env.SESSION_SECRET || 'your-secret-key',
-    resave: false,
-    saveUninitialized: false,
-    cookie: { 
-      secure: process.env.NODE_ENV === 'production',
-      maxAge: 24 * 60 * 60 * 1000 // 24 hours
-    }
-  }))
+  secret: process.env.SESSION_SECRET || 'your-secret-key',
+  resave: false,
+  saveUninitialized: false,
+  cookie: {
+    secure: process.env.NODE_ENV === 'production',
+    maxAge: 24 * 60 * 60 * 1000 // 24 hours
+  }
+}))
 
 
 
@@ -41,42 +41,42 @@ app.use(passport.session())
 
 // Local Strategy with bcrypt
 passport.use(new LocalStrategy({
-    usernameField: 'email',
-    passwordField: 'password'
-  }, async (email, password, done) => {
-    try {
-      const { rows } = await db.query(
-        'SELECT * FROM "user" WHERE email = $1',
-        [email]
-      );
-      
-      if (!rows.length) return done(null, false);
-      
-      const user = rows[0];
-      const isValid = await bcrypt.compare(password, user.password);
-      
-      if (!isValid) return done(null, false);
-      
-      return done(null, { id: user.id, email: user.email, isadmin: user.isadmin ,name: user.name});
-    } catch (error) {
-      return done(error);
-    }
-  }));
-  
-  // Serialization/Deserialization
-  passport.serializeUser((user, done) => done(null, user.id));
-  passport.deserializeUser(async (id, done) => {
-    try {
-      const { rows } = await db.query(
-        'SELECT id, email , isadmin ,name FROM "user" WHERE id = $1',
-        [id]
-      );
-      //console.log("test values" , rows[0])
-      done(null, rows[0]);
-    } catch (error) {
-      done(error);
-    }
-  });
+  usernameField: 'email',
+  passwordField: 'password'
+}, async (email, password, done) => {
+  try {
+    const { rows } = await db.query(
+      'SELECT * FROM "user" WHERE email = $1',
+      [email]
+    );
+
+    if (!rows.length) return done(null, false);
+
+    const user = rows[0];
+    const isValid = await bcrypt.compare(password, user.password);
+
+    if (!isValid) return done(null, false);
+
+    return done(null, { id: user.id, email: user.email, isadmin: user.isadmin, name: user.name });
+  } catch (error) {
+    return done(error);
+  }
+}));
+
+// Serialization/Deserialization
+passport.serializeUser((user, done) => done(null, user.id));
+passport.deserializeUser(async (id, done) => {
+  try {
+    const { rows } = await db.query(
+      'SELECT id, email , isadmin ,name FROM "user" WHERE id = $1',
+      [id]
+    );
+    //console.log("test values" , rows[0])
+    done(null, rows[0]);
+  } catch (error) {
+    done(error);
+  }
+});
 
 /*app.use((req, res, next) => {
     // Assuming `req.user` is set by passport or some authentication middleware
@@ -88,31 +88,31 @@ passport.use(new LocalStrategy({
 
 
 
- 
+
 // cross to prepare communicate with client server (React)
 app.use(cors({
-    origin: 'http://localhost:5173', //React link. we have to check if will work normally or not
-    credentials: true
-  }));
+  origin: 'http://localhost:5173', //React link. we have to check if will work normally or not
+  credentials: true
+}));
 
- 
+
 
 db.on('error', error => {
-    console.log("\x1b[31m%s\x1b[0m", "[ DB problem ]")
-    console.log(error)
+  console.log("\x1b[31m%s\x1b[0m", "[ DB problem ]")
+  console.log(error)
 })
 
 
 app.get("/", (req, res) => {
-    res.json("Home page :) ")
+  res.json("Home page :) ")
 })
 
-// routers import:
-import mainRouter from './routers/mainRouter.js' // one router for all routes n need any route in index.js file
+// routes import:
+import mainRouter from './routes/mainRouter.js' // one router for all routes n need any route in index.js file
 
 
-// routers middlewares
-app.use("/api",mainRouter)
+// routes middlewares
+app.use("/api", mainRouter)
 
 // Error handling
 app.use(errorHandler);
